+++ R/currents.R
@@ -85,7 +85,8 @@ currentsfiles <- function(time.resolution = c("daily", "weekly"), ...) {
 ##' components. Multiple dates can be returned for magnitude or
 ##' direction, U or V only.
 ##'
-##' This is the "SSALTO/DUACS - DT Geostrophic Velocities - Up-to-date Global Processing". See References.
+##' This is the "DT merged all satellites Global Ocean Gridded SSALTO/DUACS Sea Surface Height L4 product and derived variables"
+##'  See References.
 ##'
 ##' @param date date or dates of data to read, see Details
 ##' @param time.resolution time resolution to read
@@ -105,35 +106,32 @@ currentsfiles <- function(time.resolution = c("daily", "weekly"), ...) {
 ##' view \[0, 360\], the default behaviour is to reset this to Atlantic
 ##' view \[-180, 180\] with \code{lon180}. 
 ##'
-##' For weekly files, this source is no longer provided by Aviso but is cached by the raadtools system at AAD/ACE, for the period 1992 to 2014. 
-##' Note that the projection was Mercator for this earlier product. 
 ##' @return \code{\link[raster]{raster}} object with the "U"
-##' (horizontal/X) and "V" (vertical/Y) components of velocity in
-##' cm/s. Setting either of the (mutually exclusive) \code{magonly}
-##' and \code{dironly} arguments returns the magnitude (in cm) or
+##' (meridional/horizontal/X) and "V" (zonal/vertical/Y) components of velocity in
+##' m/s. Setting either of the (mutually exclusive) \code{magonly}
+##' and \code{dironly} arguments returns the magnitude (in m/s) or
 ##' direction (in degrees relative to North) of the velocity vectors.
 ##' @seealso \code{\link{icefiles}} for details on the repository of
 ##' data files, \code{\link[raster]{raster}} for the return value
 # imports should not be necessary here
 ##' @importFrom raster t flip atan2
 ##' @export
-##' @references \url{http://www.aviso.oceanobs.com/en/data/products/sea-surface-height-products/global/index.html}
+##' @references \url{http://marine.copernicus.eu}
 ##' @examples
-##' \dontrun{
-##' ## read a single time slice, and plot the directions [0,360) as an image with arrows
-##' x <- readcurr(dironly = TRUE)
-##' ## get a local extent for a zoom plot
-##' e <- extent(projectExtent(raster(extent(130, 150, -50, -30), crs = "+proj=longlat"), projection(x)))
-##' x1 <- crop(readcurr(), e)
-##' crds <- coordinates(x1)
-##' scale <- 2000
-##' plot(crop(x, e))
-##' x1 <- crds[,1]
-##' y1 <- crds[,2]
-##' x2 <- crds[,1] + values(x1[["U"]]) * scale
-##' y2 <- crds[,2] + values(x1[["V"]]) * scale
-##' arrows(x1, y1, x2, y2, length = 0.03)
-##' }
+#' ## read a single time slice, and plot the directions [0,360) as an image with arrows
+#' x <- readcurr(dironly = TRUE)
+#' ## get a local extent for a zoom plot
+#' e <- extent(projectExtent(raster(extent(130, 150, -50, -30), crs = "+proj=longlat"), projection(x)))
+#' x <- crop(readcurr(), e)
+#' crds <- coordinates(x)
+#' scale <- 1.5
+#' vlen <- function(x) sqrt(x[[1]]^2 + x[[2]]^2)
+#' plot(vlen(crop(x, e)))
+#' x1 <- crds[,1]
+#' y1 <- crds[,2]
+#' x2 <- crds[,1] + values(x[[1]]) * scale
+#' y2 <- crds[,2] + values(x[[1]]) * scale
+#' arrows(x1, y1, x2, y2, length = 0.03)
 readcurr <- function (date, time.resolution = c("daily", "weekly"),
                       xylim = NULL, lon180 = TRUE, 
                       magonly = FALSE,